Development Update 2
General:
-replaced usage of float data types with double for increased accuracy
main:
-updated importing of energetic disorder parameters
Polaron class:
-updated constructor function to also require the polaron charge as an
input variable
-added getCharge function to access the charge of the Polaron
OSC_Sim class:
-added homo and lumo energy levels for the donor and acceptor sites
-added separate energetic disroder parameters for the donor and acceptor
sites
-added separate site energy vectors for the donor and acceptor sites
-added a number of function declarations for needed functions that will
be added later
-replaced separate getElectronIt and getHoleIt by a more general
function getPolaronIt
-added getSiteType function to return the type of the site at the input
coords
-simplified calculateExciton events function
-implemented calculatePolaronEvents function
-updated executeExcitonCreation function to also add a placeholder
exciton dissociation event to the exciton dissociation events list
-updated executeNextEvent to detect all currently implemented events
-simplified getExcitonIt function
-finished implementing initializeArchitecture function
-added function siteContainsHole to check whether the site at the input
coords is occupied by a hole polaron

+++ Exciton.h
@@ -31,7 +31,7 @@ class Exciton_Creation : public Event{
         void calculateEvent(const Coords& dest_coords,const double distance,const double E_delta,const int temperature, const double prefactor){
             // No destination coords.  Destination coords are chosen upon execution.
             // No target object
-            setWaitTime((-1/prefactor)*log((float)rand01()));
+            setWaitTime((-1/prefactor)*log(rand01()));
         }
         string getName(){return name;}
     private:
@@ -45,7 +45,7 @@ class Exciton_Hop : public Event{
         void calculateEvent(const Coords& dest_coords,const double distance,const double E_delta,const int temperature, const double prefactor){
             setDestCoords(dest_coords);
             // No target object
-            setWaitTime((-1/(prefactor*intpow(1/(float)distance,6)*exp(-(float)E_delta/(K_b*temperature))))*log((float)rand01()));
+            setWaitTime((-1/(prefactor*intpow(1/distance,6)*exp(-E_delta/(K_b*temperature))))*log(rand01()));
         }
         string getName(){return name;}
     private:
@@ -58,7 +58,7 @@ class Exciton_Recombination : public Event{
         void calculateEvent(const Coords& dest_coords,const double distance,const double E_delta,const int temperature, const double prefactor){
             setDestCoords(dest_coords);
             // No target object
-            setWaitTime(-1*prefactor*log((float)rand01()));
+            setWaitTime(-1*prefactor*log(rand01()));
         }
         string getName(){return name;}
     private:
